Set meson warning_level and fix issues.
Meson makes it pretty easy to set warnings consistently across
compilers via the warning level.  For gcc and clang warning_level=3 sets
-Wall, -Wextra, and -Wpedantic.

I tried the next level which does -Weverything but that caused a lot of
issues.  There were a lot of warning found in the dependencies
(refl-cpp and fmt especially).  I also had a lot of false positive
errors.  For example, a warning staying that the code was implicitly
casting between types when it was actually explicitly casting.

I fixed all but one of the issues that were found by the compiler.

+++ test/qa_scalar.cpp
@@ -162,7 +162,7 @@ TYPED_TEST(PmtScalarFixture, explicit_cast)
     // Cast up to double if possible
     if constexpr (!Complex<TypeParam>) {
         auto z = pmtv::cast<double>(x);
-        EXPECT_TRUE(this->get_value() == z);
+        EXPECT_TRUE(double(this->get_value()) == z);
     }
 
     // Fail if we try to get a container type
